requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
django-dotenv==1.4.2
djangorestframework==3.14.0
psycopg2==2.9.10
pytz==2023.3
sqlparse==0.5.3
tzdata==2023.3
